requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-redis==5.4.0
django-unfold==0.40.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
celery[redis]==5.4.0
flower==2.0.1
psycopg2==2.9.9